adds import file item normalizer and router to normalize
