+++ dlt/normalize/items_normalizers.py
@@ -407,3 +407,25 @@ class ArrowItemsNormalizer(ItemsNormalizer):
         )
 
         return base_schema_update
+
+
+class FileImportNormalizer(ItemsNormalizer):
+    def __call__(self, extracted_items_file: str, root_table_name: str) -> List[TSchemaUpdate]:
+        logger.info(
+            f"Table {root_table_name} {self.item_storage.writer_spec.file_format} file"
+            f" {extracted_items_file} will be directly imported without normalization"
+        )
+        with self.normalize_storage.extracted_packages.storage.open_file(
+            extracted_items_file, "rb"
+        ) as f:
+            # TODO: sniff the schema depending on a file type
+            file_metrics = DataWriterMetrics(extracted_items_file, 0, f.tell(), 0, 0)
+        parts = ParsedLoadJobFileName.parse(extracted_items_file)
+        self.item_storage.import_items_file(
+            self.load_id,
+            self.schema.name,
+            parts.table_name,
+            self.normalize_storage.extracted_packages.storage.make_full_path(extracted_items_file),
+            file_metrics,
+        )
+        return []
